add ResultAnnouncement for WinActivity
